# Faster vpImage<vpRGBa> to BGR cv::Mat conversion