|                     | 640x360 (old) | 640x360 (new) | x-factor | 1280x720 (old) | 1280x720 (new) | x-factor
| --------------   | -------------        | -------------         | ---------- | -------------           | ----------------------- | ----------
| Mean (ms)   | 0.760124          | 0.258162            | 2.94      | 2.16438             | 0.671958                | 3.22
| Median(ms) | 0.75 ms          | 0.215088          |              | 2.15796                | 0.566895                |
| Std (ms)      | 0.145263        | 0.117033          |              | 0.393432              | 0.215827              |